fix(eve): keep cancellation helper workflow-safe
