+++ packages/eve/src/context/dynamic-model-lifecycle.ts
@@ -90,17 +90,6 @@ export function getActiveDynamicModelSelection(ctx: {
   return null;
 }
 
-/** Keeps a completed session-scoped model selection when its turn is cancelled. */
-export function preserveSerializedSessionDynamicModelSelection(
-  original: Record<string, unknown>,
-  interrupted: Record<string, unknown>,
-): Record<string, unknown> {
-  const selection = interrupted[SessionDynamicModelReferenceKey.name];
-  return selection === undefined
-    ? original
-    : { ...original, [SessionDynamicModelReferenceKey.name]: selection };
-}
-
 export async function dispatchDynamicModelEvent(input: {
   readonly ctx: AlsContext;
   readonly dynamicModel: RuntimeDynamicModelReference | undefined;
